requirements.txt
netaddr
pyvmomi
PyYAML
argparse
